+++ src/wren_vm.c
@@ -332,7 +332,7 @@ static ObjString* methodNotFound(WrenVM* vm, ObjClass* classObj, int symbol)
   // method expects.
   const char* methodName = vm->methodNames.data[symbol];
 
-  int methodLength = strlen(methodName);
+  int methodLength = (int)strlen(methodName);
   int numParams = 0;
   while (methodName[methodLength - numParams - 1] == ' ') numParams++;
 
@@ -343,7 +343,7 @@ static ObjString* methodNotFound(WrenVM* vm, ObjClass* classObj, int symbol)
           numParams,
           numParams == 1 ? "" : "s");
 
-  return AS_STRING(wrenNewString(vm, message, strlen(message)));
+  return AS_STRING(wrenNewString(vm, message, (int)strlen(message)));
 }
 
 // Pushes [function] onto [fiber]'s callstack and invokes it. Expects [numArgs]
@@ -865,7 +865,7 @@ static bool runInterpreter(WrenVM* vm)
       if (!IS_CLASS(expected))
       {
         const char* message = "Right operand must be a class.";
-        RUNTIME_ERROR(AS_STRING(wrenNewString(vm, message, strlen(message))));
+        RUNTIME_ERROR(AS_STRING(wrenNewString(vm, message, (int)strlen(message))));
       }
 
       ObjClass* actual = wrenGetClass(vm, POP());
@@ -997,7 +997,7 @@ static bool runInterpreter(WrenVM* vm)
             "Class '%s' may not have more than %d fields, including inherited "
             "ones.", name->value, MAX_FIELDS);
 
-        RUNTIME_ERROR(AS_STRING(wrenNewString(vm, message, strlen(message))));
+        RUNTIME_ERROR(AS_STRING(wrenNewString(vm, message, (int)strlen(message))));
       }
 
       PUSH(OBJ_VAL(classObj));
@@ -1107,9 +1107,9 @@ static void defineMethod(WrenVM* vm, const char* className,
 {
   ASSERT(className != NULL, "Must provide class name.");
 
-  int length = strlen(methodName);
+  int length = (int)strlen(methodName);
   ASSERT(methodName != NULL, "Must provide method name.");
-  ASSERT(strlen(methodName) < MAX_METHOD_NAME, "Method name too long.");
+  ASSERT((int)strlen(methodName) < MAX_METHOD_NAME, "Method name too long.");
 
   ASSERT(numParams >= 0, "numParams cannot be negative.");
   ASSERT(numParams <= MAX_PARAMETERS, "Too many parameters.");
@@ -1118,7 +1118,7 @@ static void defineMethod(WrenVM* vm, const char* className,
 
   // Find or create the class to bind the method to.
   int classSymbol = wrenSymbolTableFind(&vm->globalNames,
-                               className, strlen(className));
+                               className, (int)strlen(className));
   ObjClass* classObj;
 
   if (classSymbol != -1)
@@ -1129,7 +1129,7 @@ static void defineMethod(WrenVM* vm, const char* className,
   else
   {
     // The class doesn't already exist, so create it.
-    int length = strlen(className);
+    int length = (int)strlen(className);
     ObjString* nameString = AS_STRING(wrenNewString(vm, className, length));
 
     wrenPushRoot(vm, (Obj*)nameString);
@@ -1231,7 +1231,7 @@ void wrenReturnString(WrenVM* vm, const char* text, int length)
   ASSERT(text != NULL, "String cannot be NULL.");
 
   int size = length;
-  if (length == -1) size = strlen(text);
+  if (length == -1) size = (int)strlen(text);
 
   *vm->foreignCallSlot = wrenNewString(vm, text, size);
   vm->foreignCallSlot = NULL;
